Synchronize text view and outliner row heights

+++ OutlinerRowModel.swift
@@ -13,5 +13,5 @@ struct OutlinerRowModel: Identifiable {
 	var depth: Int
 	var isExpanded: Bool
 	var hasChildren: Bool
-	var height: Int
+	var height: CGFloat
 }
